requirements.txt
streamlit
pandas
textblob
transformers
streamlit-authenticator
pyyaml